v0.15.0 — JavaScript refactor and lint fix.
