+++ CHANGELOG.md
@@ -4,6 +4,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.15.0
+------------------------------
+*January 30, 2018*
+
+### Changed
+- Updated `.gitignore` file.
+- Added `lintPaths` to gulp config.
+- Moved the accessibility related JavaScript out into own module.
+
+
 v0.14.0
 ------------------------------
 *January 29, 2018*
