+++ gulpfile.js
@@ -1,4 +1,8 @@
 const gulp = require('gulp');
 const { build } = require('@justeat/gulp-build-fozzie');
 
-build(gulp);
+build(gulp, {
+    js: {
+        lintPaths: ['{src,test}/**/*.js']
+    }
+});
