+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@justeat/f-header",
   "description": "Fozzie Header – Header Component for Just Eat projects",
-  "version": "0.14.0",
+  "version": "0.15.0",
   "main": "dist/js/index.js",
   "files": [
     "dist",
